self updates
- Change subcommand from `upgrade` to `update`
- Use a `BufReader` for buffering file download instead of
  doing it manually with a `Vec`
- Update documentation
- Cleanup `main` module -- move `listen` stuff out to `listen`
