+++ src/main.rs
@@ -14,11 +14,11 @@ extern crate flate2;
 extern crate tar;
 
 #[macro_use] mod errors;
-mod upgrade;
+mod listen;
+mod update;
 
 use clap::{App, Arg, SubCommand, ArgMatches, AppSettings};
 use notify_rust::{Notification, Timeout};
-use chrono::Local;
 
 use std::io;
 use std::env;
@@ -39,6 +39,236 @@ pub static DEFAULT_TIMEOUT:     u32          = 10000;
 pub static DEFAULT_TIMEOUT_SECONDS_STR: &'static str = "10";
 
 
+/// Notification information to send over the wire from a remote client
+/// to a local listening server
+#[derive(Debug, Serialize, Deserialize)]
+pub struct ApiNote {
+    title: String,
+    msg: String,
+    timeout: u32,
+}
+impl ApiNote {
+    /// Create a new api-note with a message and default values
+    fn with_msg(msg: &str) -> ApiNote {
+        ApiNote {
+            title: DEFAULT_TITLE.to_owned(),
+            msg: msg.to_owned(),
+            timeout: DEFAULT_TIMEOUT,
+        }
+    }
+
+    /// Set a title, overriding the default
+    fn title(mut self, title: &str) -> ApiNote {
+        self.title = title.to_owned();
+        self
+    }
+
+    /// Set a timeout in milliseconds, overriding the default
+    fn timeout(mut self, millis: u32) -> ApiNote {
+        self.timeout = millis;
+        self
+    }
+}
+
+
+/// Notification builder
+pub struct Note {
+    title: String,
+    msg: String,
+    timeout: u32,
+    send: bool,
+    host: String,
+    port: u32,
+}
+impl Note {
+    /// Create a new notification with a given message and default values
+    pub fn with_msg(msg: &str) -> Note {
+        Note {
+            title: DEFAULT_TITLE.to_owned(),
+            msg: msg.to_owned(),
+            timeout: DEFAULT_TIMEOUT,
+            send: false,
+            host: DEFAULT_HOST.to_owned(),
+            port: DEFAULT_PORT }
+    }
+
+    /// Set a title, overriding the default
+    pub fn title(mut self, title: &str) -> Note {
+        self.title = title.to_owned();
+        self
+    }
+
+    /// Set timeout in milliseconds, overriding the default
+    pub fn timeout(mut self, millis: u32) -> Note {
+        self.timeout = millis;
+        self
+    }
+
+    /// Set whether the notification should send itself to a listener
+    pub fn send(mut self, send: bool) -> Note {
+        self.send = send;
+        self
+    }
+
+    /// Set the receiving host (for sending), replacing the default
+    pub fn host(mut self, host: &str) -> Note {
+        self.host = host.to_owned();
+        self
+    }
+
+    /// Set the receiving port (for sending), replacing the default
+    pub fn port(mut self, port: u32) -> Note {
+        self.port = port;
+        self
+    }
+
+    /// Create the notification locally, or send it over the wire
+    ///
+    /// Errors:
+    ///     * Serializing `ApiNote`
+    ///     * Connecting to a listener
+    ///     * Writing to listener stream
+    ///     * Communicating to the system notification-server
+    pub fn push(self) -> Result<()> {
+        if self.send {
+            use io::Write;
+            let addr = format!("{}:{}", self.host, self.port);
+            let note = ApiNote::with_msg(&self.msg)
+                .title(&self.title)
+                .timeout(self.timeout);
+            let note = serde_json::to_string(&note)?;
+            let mut stream = net::TcpStream::connect(&addr)?;
+            stream.write(note.as_bytes())?;
+        } else {
+            Notification::new()
+                .icon(DEFAULT_ICON)
+                .summary(&self.title)
+                .timeout(Timeout::Milliseconds(self.timeout))
+                .body(&self.msg)
+                .show()?;
+        }
+        Ok(())
+    }
+}
+
+
+/// Check if we can connect to the specified receiver
+///
+/// Errors:
+///     * Connecting to listener
+///     * Writing to listener stream
+fn can_connect(host: &str, port: u32) -> Result<()> {
+    use io::Write;
+    let addr = format!("{}:{}", host, port);
+    let mut stream = net::TcpStream::connect(&addr)?;
+    stream.write("ping".as_bytes())?;
+    Ok(())
+}
+
+
+/// Run a command in foreground
+///
+/// Errors:
+///     * Converting `cmd` to `Cstring`
+///     * `cmd` exited with non-zero status-code
+fn run_command(cmd: &str) -> Result<()> {
+    let c_str = ffi::CString::new(cmd)?;
+    let ret = unsafe {
+        let ret = libc::system(c_str.as_ptr());
+        // convert child status code to a normal code 0-255
+        libc::WEXITSTATUS(ret)
+    };
+    if ret != 0 { return Err(Error::Command(ret)) }
+    Ok(())
+}
+
+
+/// Collect all default and overridden notification parameters from argument matches,
+/// returning a captured command-string and a constructed `Note`
+///
+/// Errors:
+///     * Unable to parse input integers
+///     * Unable to connect to specified listener with provided `host`/`port`
+///     * No `command-string` provided
+fn collect_cmd_note(matches: &ArgMatches) -> Result<(String, Note)> {
+    // Capture default and overridden notification arguments
+    let send = matches.is_present("send") ||
+        env::var("CLIN_SEND").ok()
+            .and_then(|s| if s == "1" { Some(()) } else { None })
+            .is_some();
+    let fallback_host = env::var("CLIN_SEND_HOST").unwrap_or_else(|_| DEFAULT_HOST.to_string());
+    let host = matches.value_of("host")
+        .unwrap_or(&fallback_host);
+    let fallback_port = env::var("CLIN_SEND_PORT").unwrap_or_else(|_| DEFAULT_PORT_STR.to_string());
+    let port = matches.value_of("port")
+        .unwrap_or(&fallback_port)
+        .parse::<u32>()?;
+    let fallback_timeout = env::var("CLIN_TIMEOUT").unwrap_or_else(|_| DEFAULT_TIMEOUT_STR.to_string());
+    let timeout = matches.value_of("timeout")
+        .unwrap_or(&fallback_timeout)
+        .parse::<u32>()?;
+
+    // If sending, make sure specified connection works
+    if send && can_connect(&host, port).is_err() {
+        bail!(Error::Network, "Unable to connect to clin-listener at `{}:{}`", host, port)
+    }
+
+    // Capture command contents or bail out if nothing was provided
+    let cmd = match (matches.value_of("command_string"), matches.is_present("cmd")) {
+        (Some(c), _) => c.to_owned(),
+        (_, true) => {
+            // Pull out the full trailing args list...
+            // The built in parsing will strip out any '--' from the list
+            let args = env::args().collect::<Vec<_>>();
+            let ind = match args.iter().position(|item| item == "--") {
+                None => bail!(Error::Msg, "Error parsing command, no `--` delimiter found"),
+                Some(i) => i,
+            };
+            let (_, args) = args.split_at(ind + 1);
+            args.join(" ")
+        }
+        _ => {
+            bail_help!()
+        }
+    };
+    let note = Note::with_msg(&cmd)
+        .timeout(timeout)
+        .send(send)
+        .host(host)
+        .port(port);
+    Ok((cmd, note))
+}
+
+
+/// Dispatch over arguments
+fn run(matches: ArgMatches) -> Result<()> {
+    if let Some(self_matches) = matches.subcommand_matches("self") {
+        if let Some(update_matches) = self_matches.subcommand_matches("update") {
+            let show_progress = !update_matches.is_present("quiet");
+            return update::run(show_progress);
+        }
+        bail_help!()
+    }
+
+    // Initialize a clin-listener
+    if let Some(listen_matches) = matches.subcommand_matches("listen") {
+        return listen::start_listener(listen_matches);
+    }
+
+    let (cmd, note) = collect_cmd_note(&matches)?;
+    println!("clin: `{}`", cmd);
+
+    let title = match run_command(&cmd) {
+        Err(Error::Command(ret)) => {
+            format!("Error ✗ -- exit status: {}", ret)
+        }
+        Err(e) => return Err(e),
+        Ok(_) => "Complete ✓".to_string(),
+    };
+    note.title(&title).push()
+}
+
+
 fn main() {
     let matches = App::new("CLIN")
         .setting(AppSettings::TrailingVarArg)
@@ -53,10 +283,10 @@ clin -- ./some-build-script.sh --flag --arg1 'some arg'
 clin -c \"./some-build-script.sh --flag --arg1 'some arg'\"")
         .subcommand(SubCommand::with_name("self")
                     .about("Self referential things")
-                    .subcommand(SubCommand::with_name("upgrade")
-                        .about("Upgrade to the latest binary release, replacing this binary")
+                    .subcommand(SubCommand::with_name("update")
+                        .about("Update to the latest binary release, replacing this binary")
                         .arg(Arg::with_name("quiet")
-                             .help("Suppress unnecessary download output")
+                             .help("Suppress unnecessary download output (progress bar)")
                              .long("quiet")
                              .short("q")
                              .required(false)
@@ -124,248 +354,3 @@ clin -c \"./some-build-script.sh --flag --arg1 'some arg'\"")
     }
 }
 
-
-fn run(matches: ArgMatches) -> Result<()> {
-    if let Some(self_matches) = matches.subcommand_matches("self") {
-        if let Some(upgrade_matches) = self_matches.subcommand_matches("upgrade") {
-            let show_progress = !upgrade_matches.is_present("quiet");
-            return upgrade::run(show_progress);
-        }
-        return see_help();
-    }
-
-    // Initialize a clin-listener
-    if let Some(listen_matches) = matches.subcommand_matches("listen") {
-        init_logger(listen_matches.is_present("log"));
-        let host = if listen_matches.is_present("public") { "0.0.0.0" } else { "127.0.0.1" };
-        let fallback_port = env::var("CLIN_LISTEN_PORT").unwrap_or_else(|_| DEFAULT_PORT_STR.to_string());
-        let port = listen_matches.value_of("port")
-            .unwrap_or(&fallback_port)
-            .parse::<u32>()?;
-        let addr = format!("{}:{}", host, port);
-        return listen(&addr);
-    }
-
-    // Capture default and overridden notification arguments
-    let send = matches.is_present("send") ||
-        env::var("CLIN_SEND").ok()
-            .and_then(|s| if s == "1" { Some(()) } else { None })
-            .is_some();
-    let fallback_host = env::var("CLIN_SEND_HOST").unwrap_or_else(|_| DEFAULT_HOST.to_string());
-    let host = matches.value_of("host")
-        .unwrap_or(&fallback_host);
-    let fallback_port = env::var("CLIN_SEND_PORT").unwrap_or_else(|_| DEFAULT_PORT_STR.to_string());
-    let port = matches.value_of("port")
-        .unwrap_or(&fallback_port)
-        .parse::<u32>()?;
-    let fallback_timeout = env::var("CLIN_TIMEOUT").unwrap_or_else(|_| DEFAULT_TIMEOUT_STR.to_string());
-    let timeout = matches.value_of("timeout")
-        .unwrap_or(&fallback_timeout)
-        .parse::<u32>()?;
-
-    // If sending, make sure specified connection works
-    if send && can_connect(&host, port).is_err() {
-        bail!(Error::Network, "Unable to connect to clin-listener at `{}:{}`", host, port)
-    }
-
-    // Capture command contents or bail out if nothing was provided
-    let cmd = match (matches.value_of("command_string"), matches.is_present("cmd")) {
-        (Some(c), _) => c.to_owned(),
-        (_, true) => {
-            // Pull out the full trailing args list...
-            // The built in parsing will strip out any '--' from the list
-            let args = env::args().collect::<Vec<_>>();
-            let ind = match args.iter().position(|item| item == "--") {
-                None => bail!(Error::Msg, "Error parsing command, no `--` delimiter found"),
-                Some(i) => i,
-            };
-            let (_, args) = args.split_at(ind + 1);
-            args.join(" ")
-        }
-        _ => {
-            return see_help();
-        }
-    };
-
-    println!("clin: `{}`", cmd);
-
-    let title = match run_command(&cmd) {
-        Err(Error::Command(ret)) => {
-            format!("Error ✗ -- exit status: {}", ret)
-        }
-        Err(e) => return Err(e),
-        Ok(_) => "Complete ✓".to_string(),
-    };
-
-    Note::with_msg(&cmd)
-        .title(&title)
-        .timeout(timeout)
-        .send(send)
-        .host(host)
-        .port(port)
-        .push()
-}
-
-
-fn see_help() -> Result<()> {
-    println!("clin: see `--help`");
-    return Ok(());
-}
-
-
-/// Initialize loggers for the listening server
-fn init_logger(log: bool) {
-    if log {
-        env::set_var("LOG", "info")
-    }
-
-    // Set a custom logging format & change the env-var to "LOG"
-    // e.g. LOG=info clin listen
-    env_logger::LogBuilder::new()
-        .format(|record| {
-            format!("{} [{}] - [{}] -> {}",
-                Local::now().format("%Y-%m-%d_%H:%M:%S"),
-                record.level(),
-                record.location().module_path(),
-                record.args()
-                )
-            })
-        .parse(&env::var("LOG").unwrap_or_default())
-        .init()
-        .expect("failed to initialize logger");
-}
-
-
-/// Check if we can connect to the specified receiver
-fn can_connect(host: &str, port: u32) -> Result<()> {
-    use io::Write;
-    let addr = format!("{}:{}", host, port);
-    let mut stream = net::TcpStream::connect(&addr)?;
-    stream.write("ping".as_bytes())?;
-    Ok(())
-}
-
-
-/// Listen on the given address for incoming `ApiNote` messages
-/// and generate local notifications
-fn listen(addr: &str) -> Result<()> {
-    use io::Read;
-    info!("** Listening on {} **", addr);
-
-    let listener = net::TcpListener::bind(&addr)?;
-    for stream in listener.incoming() {
-        let mut stream = stream?;
-        let mut s = String::new();
-        stream.read_to_string(&mut s)?;
-        if s == "ping" { continue; }
-        let note: ApiNote = serde_json::from_str(&s)?;
-        info!("[{}]: {}", note.title, note.msg);
-        Note::with_msg(&note.msg)
-            .title(&note.title)
-            .timeout(note.timeout)
-            .push()?;
-    }
-    Ok(())
-}
-
-
-/// Run a command in foreground
-fn run_command(cmd: &str) -> Result<()> {
-    let c_str = ffi::CString::new(cmd)?;
-    let ret = unsafe {
-        let ret = libc::system(c_str.as_ptr());
-        // convert child status code to a normal code 0-255
-        libc::WEXITSTATUS(ret)
-    };
-    if ret != 0 { return Err(Error::Command(ret)) }
-    Ok(())
-}
-
-
-/// Notification information to send over the wire from a remote client
-/// to a local listening server
-#[derive(Debug, Serialize, Deserialize)]
-struct ApiNote {
-    title: String,
-    msg: String,
-    timeout: u32,
-}
-impl ApiNote {
-    fn with_msg(msg: &str) -> ApiNote {
-        ApiNote {
-            title: DEFAULT_TITLE.to_owned(),
-            msg: msg.to_owned(),
-            timeout: DEFAULT_TIMEOUT,
-        }
-    }
-    fn title(mut self, title: &str) -> ApiNote {
-        self.title = title.to_owned();
-        self
-    }
-    fn timeout(mut self, millis: u32) -> ApiNote {
-        self.timeout = millis;
-        self
-    }
-}
-
-
-/// Notification builder
-struct Note {
-    title: String,
-    msg: String,
-    timeout: u32,
-    send: bool,
-    host: String,
-    port: u32,
-}
-impl Note {
-    fn with_msg(msg: &str) -> Note {
-        Note {
-            title: DEFAULT_TITLE.to_owned(),
-            msg: msg.to_owned(),
-            timeout: DEFAULT_TIMEOUT,
-            send: false,
-            host: DEFAULT_HOST.to_owned(),
-            port: DEFAULT_PORT }
-    }
-    fn title(mut self, title: &str) -> Note {
-        self.title = title.to_owned();
-        self
-    }
-    fn timeout(mut self, millis: u32) -> Note {
-        self.timeout = millis;
-        self
-    }
-    fn send(mut self, send: bool) -> Note {
-        self.send = send;
-        self
-    }
-    fn host(mut self, host: &str) -> Note {
-        self.host = host.to_owned();
-        self
-    }
-    fn port(mut self, port: u32) -> Note {
-        self.port = port;
-        self
-    }
-    fn push(self) -> Result<()> {
-        if self.send {
-            use io::Write;
-            let addr = format!("{}:{}", self.host, self.port);
-            let note = ApiNote::with_msg(&self.msg)
-                .title(&self.title)
-                .timeout(self.timeout);
-            let note = serde_json::to_string(&note)?;
-            let mut stream = net::TcpStream::connect(&addr)?;
-            stream.write(note.as_bytes())?;
-        } else {
-            Notification::new()
-                .icon(DEFAULT_ICON)
-                .summary(&self.title)
-                .timeout(Timeout::Milliseconds(self.timeout))
-                .body(&self.msg)
-                .show()?;
-        }
-        Ok(())
-    }
-}
